Load data from snapshot if available
If the snapshot feature is configured it a snapshot is loaded to see
the resource configuration table. This means that the snapshot can be
used to serve connecting Envoy instances, even if not configuration
can be generated by the configured adapter.

This makes the ExControlPlane more robust against bad updates of the
adapter and / or inputs.